Add styles folder
