+++ CourseAPP/Components/Courses.js
@@ -3,6 +3,7 @@ import { View, Text, ScrollView, RefreshControl, TouchableOpacity, StyleSheet } 
 import axios from 'axios';
 import { useNavigation, useFocusEffect } from '@react-navigation/native';
 import config from './ConfigIp';
+import styles from './Styles/CoursesStyles'; 
 
 const Courses = () => {
   const [courses, setCourses] = useState([]);
@@ -50,7 +51,7 @@ const Courses = () => {
     }, [])
   );
 
-  return (
+ return (
     <View style={styles.container}>
       <ScrollView
         contentContainerStyle={styles.scrollContainer}
@@ -86,7 +87,6 @@ const Courses = () => {
         )}
       </ScrollView>
 
-      {/* Botón fijo */}
       <TouchableOpacity
         onPress={() => navigation.navigate('AddCourse')}
         style={styles.addButton}
@@ -96,67 +96,5 @@ const Courses = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1, // Muy importante para ocupar toda la pantalla
-  },
-  scrollContainer: {
-    padding: 10,
-    paddingBottom: 80, // Para que el último curso no quede tapado por el botón
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginVertical: 10,
-    textAlign: 'center',
-  },
-  courseItem: {
-    marginBottom: 15,
-    padding: 15,
-    backgroundColor: '#f9f9f9',
-    borderRadius: 10,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  courseName: {
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  courseDescription: {
-    fontSize: 14,
-    marginVertical: 5,
-  },
-  buttonsContainer: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
-  button: {
-    padding: 10,
-    borderRadius: 5,
-  },
-  buttonText: {
-    color: 'white',
-  },
- addButton: {
-  backgroundColor: 'green',
-  padding: 15,
-  borderRadius: 5,
-  margin: 10,
-  position: 'absolute',
-  bottom: 35, // Ahora el botón está 30 unidades más arriba
-  left: 10,
-  right: 10,
-},
-
-  addButtonText: {
-    color: 'white',
-    textAlign: 'center',
-    fontWeight: 'bold',
-  },
-});
 
 export default Courses;
