+++ CourseAPP/Components/Styles/AddCourseStyles.js
@@ -1,0 +1,37 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#f9f9f9',
+    padding: 20,
+  },
+  input: {
+    height: 50,
+    borderColor: '#ddd',
+    borderWidth: 1,
+    borderRadius: 10,
+    paddingHorizontal: 15,
+    marginBottom: 15,
+    backgroundColor: '#fff',
+    fontSize: 16,
+  },
+  textArea: {
+    height: 100,
+    textAlignVertical: 'top', // Para que el texto comience desde arriba
+  },
+  buttonContainer: {
+    marginTop: 10,
+    backgroundColor: '#007bff',
+    borderRadius: 10,
+  },
+  buttonText: {
+    color: '#fff',
+    textAlign: 'center',
+    paddingVertical: 12,
+    fontSize: 16,
+    fontWeight: '600',
+  },
+});
+
+export default styles;
