+++ CourseAPP/Components/Styles/CoursesStyles.js
@@ -1,0 +1,64 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  scrollContainer: {
+    padding: 10,
+    paddingBottom: 80,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginVertical: 10,
+    textAlign: 'center',
+  },
+  courseItem: {
+    marginBottom: 15,
+    padding: 15,
+    backgroundColor: '#f9f9f9',
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  courseName: {
+    fontSize: 18,
+    fontWeight: '600',
+  },
+  courseDescription: {
+    fontSize: 14,
+    marginVertical: 5,
+  },
+  buttonsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+  },
+  button: {
+    padding: 10,
+    borderRadius: 5,
+  },
+  buttonText: {
+    color: 'white',
+  },
+  addButton: {
+    backgroundColor: 'green',
+    padding: 15,
+    borderRadius: 5,
+    margin: 10,
+    position: 'absolute',
+    bottom: 35,
+    left: 10,
+    right: 10,
+  },
+  addButtonText: {
+    color: 'white',
+    textAlign: 'center',
+    fontWeight: 'bold',
+  },
+});
+
+export default styles;
